+++ packages/slate-plugins/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.59.2 (2020-06-12)
+
+**Note:** Version bump only for package @udecode/slate-plugins
+
+
+
+
+
+## 0.59.1 (2020-06-06)
+
+**Note:** Version bump only for package @udecode/slate-plugins
+
+
+
+
+
 # 0.59.0 (2020-06-05)
 
 **Note:** Version bump only for package @udecode/slate-plugins